+++ src/pages/DashBoardPage.js
@@ -1,18 +1,26 @@
-import { Outlet } from 'react-router-dom';
+import { Outlet, useLocation } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
+import { isOpen } from '../redux/currency/currency-selectors';
 import { Balance } from '../components/Balance/Balance';
 import { Currency } from '../components/Currency/Currency';
 import { Navigation } from '../components/Navigation/Navigation';
 import { Header } from '../components/Header/Header';
-
 import {
   ContainerPages,
   BlurContainer,
 } from '../stylesheet/WrapperPages.styled';
+import useMediaQuery from '../hooks/useMediaQuery';
 
 export default function DashBoardPage() {
+  const { pathname } = useLocation();
+  const isMobile = useMediaQuery('(max-width: 768px)');
+  const isCurrencyOpen = useSelector(isOpen);
+  const isCurrencyShow = isMobile ? isCurrencyOpen : true;
+  const isBalanceShow = isMobile ? pathname === '/home' : true;
+
   return (
     <BlurContainer>
       <ToastContainer />
@@ -20,12 +28,14 @@ export default function DashBoardPage() {
       <ContainerPages>
         <div>
           <Navigation />
-          <Balance />
-          <Currency />
+          {isBalanceShow && <Balance />}
+          {isCurrencyShow && <Currency />}
         </div>
-        <div>
-          <Outlet />
-        </div>
+        {(!isCurrencyOpen || !isMobile) && (
+          <div>
+            <Outlet />
+          </div>
+        )}
       </ContainerPages>
     </BlurContainer>
   );
